Fixed a major issue with how libraries are created in 5.2, see below for explanation.
So open_enums() was expecting the top of the stack to be the global table "directfb" so it could
insert a bunch of tables for each enum.  This table is used by the check_<enum> functions, but
since the open_directfb() function didn't call open_enums correctly, it never gets populated, which
leads to indexing-nil errors.  Furthermore, the open_<interface> functions left an extra table on
the stack, so the directfb table would get buried.

The solution was to add a "lua_pop(L,1)" call to each open_<interface> function, and reorder
the lua_setglobal in open_directfb().

Signed-off-by: Matt Sickler <imMute@msk4.com>